refactor `paymentConfirmationPage`
