+++ lib/page/assets/transfer/paymentConfirmationPage/components/transfer_state.dart
@@ -1,0 +1,28 @@
+enum TransferState {
+  notStarted,
+  submitting,
+  finished,
+  failed,
+}
+
+extension transferStateExtension on TransferState {
+  bool isFinishedOrFailed() {
+    return this == TransferState.finished || this == TransferState.failed;
+  }
+
+  bool notStarted() {
+    return this == TransferState.notStarted;
+  }
+
+  bool isSubmitting() {
+    return this == TransferState.submitting;
+  }
+
+  bool isFinished() {
+    return this == TransferState.finished;
+  }
+
+  bool isFailed() {
+    return this == TransferState.failed;
+  }
+}
